requirements.txt
emoji==2.2.0
liquidctl==1.12.1
Pillow==9.4.0
requests==2.28.1
tweepy==4.12.1
toml==0.10.2